requirements_dev.txt
-r requirements_common.txt
selenium
django-debug-toolbar
coverage
django-jenkins
ipython
pep8
pyflakes
pylint
flake8

